giving clear calculation

+++ python_introduction/hours_to_seconds.py
@@ -1,8 +1,8 @@
 #declare variable 
 hours = 2
 
-# calculating the hours in seconds 
-seconds = int(hours * 3600)
+#calculating the hours in seconds 
+seconds = hours * 3600
 
 #print seconds in hours 
 print(f"{hours} hour(s) is {seconds} seconds")
